requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
anyascii==0.3.2
    # via wagtail
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
asttokens==2.4.1
    # via stack-data
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via wagtail
black==24.4.0
    # via -r requirements.in
brotli==1.1.0
    # via whitenoise
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via
    #   requests
    #   sentry-sdk
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
defusedxml==0.7.1
    # via willow
distlib==0.3.8
    # via virtualenv
dj-database-url==2.1.0
    # via -r requirements.in
django==5.0.4
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-anymail
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-linear-migrations
    #   django-modelcluster
    #   django-permissionedforms
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
    #   laces
    #   wagtail
django-anymail[mailgun]==10.3
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==23.5
    # via wagtail
django-linear-migrations==2.12.0
    # via -r requirements.in
django-modelcluster==6.3
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-taggit==5.0.1
    # via wagtail
django-treebeard==4.7.1
    # via wagtail
djangorestframework==3.15.1
    # via
    #   -r requirements.in
    #   djangorestframework-gis
    #   drf-spectacular
    #   wagtail
djangorestframework-gis==1.0
    # via -r requirements.in
draftjs-exporter==5.0.0
    # via wagtail
drf-spectacular[sidecar]==0.27.2
    # via -r requirements.in
drf-spectacular-sidecar==2024.4.1
    # via drf-spectacular
et-xmlfile==1.1.0
    # via openpyxl
executing==2.0.1
    # via stack-data
factory-boy==3.3.0
    # via
    #   -r requirements.in
    #   wagtail-factories
faker==24.11.0
    # via
    #   -r requirements.in
    #   factory-boy
filelock==3.13.1
    # via virtualenv
filetype==1.2.0
    # via willow
gunicorn==22.0.0
    # via -r requirements.in
html5lib==1.1
    # via wagtail
identify==2.5.33
    # via pre-commit
idna==3.6
    # via requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.0.0
    # via pytest
ipdb==0.13.13
    # via -r requirements.in
ipython==8.20.0
    # via ipdb
jedi==0.19.1
    # via ipython
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
l18n==2021.3
    # via wagtail
laces==0.1.1
    # via wagtail
matplotlib-inline==0.1.6
    # via ipython
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pre-commit
openpyxl==3.1.2
    # via wagtail
packaging==23.2
    # via
    #   black
    #   build
    #   gunicorn
    #   pytest
parso==0.8.3
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
pillow==10.3.0
    # via
    #   pillow-heif
    #   wagtail
pillow-heif==0.16.0
    # via willow
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.1.0
    # via
    #   black
    #   virtualenv
pluggy==1.4.0
    # via pytest
pre-commit==3.7.0
    # via -r requirements.in
prompt-toolkit==3.0.43
    # via ipython
psycopg[binary]==3.1.18
    # via -r requirements.in
psycopg-binary==3.1.18
    # via psycopg
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.17.2
    # via ipython
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
pytest==8.1.1
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via faker
pytz==2024.1
    # via
    #   django-modelcluster
    #   l18n
pyyaml==6.0.1
    # via
    #   drf-spectacular
    #   pre-commit
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   django-anymail
    #   wagtail
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
sentry-sdk==1.45.0
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
stack-data==0.6.3
    # via ipython
telepath==0.3.1
    # via wagtail
traitlets==5.14.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.9.0
    # via
    #   dj-database-url
    #   psycopg
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.1.0
    # via
    #   django-anymail
    #   requests
    #   sentry-sdk
virtualenv==20.25.0
    # via pre-commit
wagtail==6.0.2
    # via
    #   -r requirements.in
    #   wagtail-factories
wagtail-factories==4.1.0
    # via -r requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
wheel==0.42.0
    # via pip-tools
whitenoise[brotli]==6.6.0
    # via -r requirements.in
willow[heif]==1.8.0
    # via
    #   wagtail
    #   willow

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
